[BACKPORT 2.14.3][#14961, 14962] YSQL: Further reduce first-query metadata RPCs
Summary:
Patch on top of 2.14.3 branch for a custom build with significantly
reduced first-query overhead/latency.
The tradeoff is that the default per-connection memory usage is higher.

The core (in-progress) change included is described below:

-------
Reduce the number of catalog cache misses (and yb-master RPCs) for the first
queries in a new connection by preloading more common system catalogs and
improving our use of negative caching.

* Add pg_proc, pg_amop, pg_amproc, pg_cast, pg_operator to preloaded catalogs.
* Disable relcache init file by default and rely on tserver prefetching instead.
  Context: relcache init file only loads the relation cache, while
           tserver-prefetching includes catalog cache as well.
* Add a mechanism for better leveraging negative caching for preloaded catalogs.
  Context: for a fully preloaded catalog we can be certain that cache miss
           implies the entry really does not exist. So we can skip some
           redundant lookup for non-existent entries.
* A minor fix for incrementing catalog version for creating enum and domain
  types.
  Context: We already allowed negative caching for pg_type so incrementing the
           version was required to ensure cache invalidation on create.
           But the enum and domain type cases were missed before.

All the changes above are under a tserver gflag: `ysql_catalog_prefetech_additional_tables`
which is off (old behavior) by default.
------

Note: For this custom patch the flag is on by default instead (to run full test
suite with the flag on).

This diff also includes the changes made in the following diffs:
  - https://phabricator.dev.yugabyte.com/D18268
  - https://phabricator.dev.yugabyte.com/D18836
  - https://phabricator.dev.yugabyte.com/D18860
These diffs have not been backported to 2.14 and include changes that this backport diff depends on.

For an official release with this change/functionality we'd need to:
 1. backport the 3 changes above to 2.14
 2. Land the core change described above in master branch
 3. Change the prefetch flag default back to off in 2.14 release (to not
    affect existing users) and land the 2.14 backport

Test Plan:
Existing Jenkins tests
Manual testing with user queries and checking CatalogCacheMisses metric and
yb-master RPC counts.
Example: for a complex user-example RPC count was reduced from 484 to 28.
For a simpler query it reduced from 154 to ~10.
Tradeoff Cache memory usage goes from 12MB to 21 MB.

To enable prefetching:
`ybd create --tserver_flags "ysql_catalog_prefetch_additional_tables=true"`

Reviewers: dmitry, ssong, myang, plee

Subscribers: smishra, yql

Differential Revision: https://phabricator.dev.yugabyte.com/D21488