[SAMPLE][#TBD] YSQL: Propagate hybrid time between YSQL and local tserver proxy
Summary:
As part of commit c5f51258, a major refactor was done in 2.14 to move parts of
YSQL side client into the same node's tserver to avoid a lot of network
connections that result from multiple YSQL backends connecting to each tserver
in a cluster. This was done by creating a new PgClientService on each tserver
that would serve as a proxy for all YSQL backends of the same node.

This refactor resulted in a subtle regression where in the hybrid time wasn't
propagated from the tserver's where reads/ writes were performed to the YSQL
backend and vice-versa. The hybrid time was propagated from these tservers to
the PgClientService proxies, but not to the YSQL backends.

This results in certain scenarios where the clock skew can result in certain
data visibility issues. Details of one such visibility bug will be added here
shortly.

The diff is a sample that -
(1) adds a log line to confirm the theory around one such visibility bug
(2) has a fix for the visibility bug under a gflag. The fix is not yet generic
and is specific to that one visibility bug. A generic fix will follow in this
diff once we have validation that it works.

NOTE: this fix is created on top of 2.14.2.0 for testing purposes. Once validated
I will fix this in master first and then backport all the way to 2.14.

Test Plan: Jenkins

Subscribers: yql, bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D22819